+++ kpopbot/__init__.py
@@ -1,4 +1,6 @@
+from .utils import sanitize_flair
+from .CONSTANTS import *
 from .media import media_provider
 from .post import post_factory
 from .reddit import PRAWWrapper
-from .utils import sanitize_flair
+
