+++ kpopbot/reddit.py
@@ -1,12 +1,9 @@
-from CONFIG import REDDIT_CONFIG
-from CONSTANTS import WHITELIST
 from praw import Reddit
-from .utils import sanitize_flair as sf
 
 
 class PRAWWrapper:
     def __init__(self, config):
-        self.instance = Reddit(**REDDIT_CONFIG).subreddit("kpop")
+        self.instance = Reddit(**config).subreddit("kpop")
 
     def obtain_posts(self):
         return (
